Fix curve lib

+++ contracts/core/lib/Constants.sol
@@ -8,13 +8,13 @@ library Constants {
     string constant PREFIX_LP_XT = "LpXT:";
     string constant PREFIX_GNFT = "GNFT:";
     string constant STRING_UNDER_LINE = "_";
-    uint32 constant DECIMAL_BASE = 1e8;
-    uint64 constant DECIMAL_BASE_SQ = 1e16;
-    uint16 constant DAYS_IN_YEAR = 365;
-    uint32 constant SECONDS_IN_DAY = 86400;
-    uint32 constant SECONDS_IN_MOUNTH = 2592000;
+    uint256 constant DECIMAL_BASE = 1e8;
+    uint256 constant DECIMAL_BASE_SQ = 1e16;
+    uint256 constant DAYS_IN_YEAR = 365;
+    uint256 constant SECONDS_IN_DAY = 86400;
+    uint256 constant SECONDS_IN_MOUNTH = 2592000;
     // The percentage of repay amount to liquidator while do liquidate
-    uint32 constant REWARD_TO_LIQUIDATOR = 5e6;
+    uint256 constant REWARD_TO_LIQUIDATOR = 5e6;
     // The percentage of repay amount to protocol while do liquidate
-    uint32 constant REWARD_TO_PROTOCOL = 5e6;
+    uint256 constant REWARD_TO_PROTOCOL = 5e6;
 }
